Some initial work to support Travis
I'm going to stop here but this isn't really right. What I'd like to do
is setup Travis and deploy to Heroku and then see if this works the way
I'm expecting. The thing that's sort tricky here is that I need to see
some hooks from Travis before I'm sure that I know how to parse them.