Add a way to suppress hooks
Some of these hooks need to be suppressed. For example. When you push to
delete a branch on GitHub, that results in two hooks: one for the push
and another for the branch deletion. What I'm going to do in that case
is suppress the push hook, but there are other cases. Like the ping
events from GitHub.