Add Projects to the model-layer
The idea here is that in order to support multiple users, I'm going to
have to have a join model called Subscription that connects a User to
the Hooks they care about. This is the first major restructuring commit
that gets me there.

By inserting a model into the relationships here, I've had to make a lot
of code changes, but mostly to tests! I also took this opportunity to
put some specs around the parsers - it aided in refactoring and gave me
more confidence that my refactor was done correctly.

Finally, this commit does change the API, instead of using `project` as
the key for the name of the project, I'm using `project_name`. This made
sense to me because it keeps Rails simple and is more correct - the
value here is not an object, it's a string and the key name of
`project_name` conveys that difference.